+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added default values for `umls_api_key` and `conn_fun` 
 arguments  
+* Fixed bug with `prepare_cpt4` where processed CONCEPT.csv 
+is replaced with the raw CONCEPT.csv  
 
 
 # setupAthena 0.7.0 (2021-12-09)